requirements.txt
PyYAML>=6.0
jsonpath-ng>=1.7.0
deepdiff>=8.4.2